fix housing object init structure, game no longer crashes/doesn't load

+++ sql/schema/schema.sql
@@ -579,7 +579,7 @@ CREATE TABLE `landplaceditems` (
 	`PosX` FLOAT NOT NULL,
 	`PosY` FLOAT NOT NULL,
 	`PosZ` FLOAT NOT NULL,
-	`Rotation` INT(10) NOT NULL,
+	`Rotation` FLOAT NOT NULL,
 	PRIMARY KEY (`ItemId`)
 )
 COLLATE='latin1_swedish_ci' ENGINE=InnoDB;
